fix some bugs

+++ IMS-backend/src/main/java/com/example/imsbackend/service/impl/SchoolServiceImpl.java
@@ -1,18 +1,24 @@
 package com.example.imsbackend.service.impl;
 
+import cn.hutool.core.util.ObjectUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.example.imsbackend.entity.School;
+import com.example.imsbackend.entity.UserSchool;
+import com.example.imsbackend.handler.exception.SchoolDeleteException;
 import com.example.imsbackend.mapper.SchoolMapper;
+import com.example.imsbackend.mapper.UserSchoolMapper;
 import com.example.imsbackend.service.SchoolService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 
 @Service("schoolService")
+@RequiredArgsConstructor
 public class SchoolServiceImpl extends ServiceImpl<SchoolMapper, School> implements SchoolService {
-
+    private final UserSchoolMapper userSchoolMapper;
     @Override
     public List<School> getAllSchool() {
         return baseMapper.selectList(new LambdaQueryWrapper<>());
@@ -30,6 +36,12 @@ public class SchoolServiceImpl extends ServiceImpl<SchoolMapper, School> impleme
 
     @Override
     public boolean deleteSchoolById(Integer id) {
-        return baseMapper.deleteById(id) == 1;
+        UserSchool userSchool = userSchoolMapper.selectOne(new LambdaQueryWrapper<UserSchool>()
+                .eq(UserSchool::getSchoolId, id));
+        if(ObjectUtil.isEmpty(userSchool)){
+            return baseMapper.deleteById(id) == 1;
+        } else {
+            throw new SchoolDeleteException();
+        }
     }
 }
